add Filters, PassFilter and BinaryOperation modules
test clone and equals methods for Biquad and PassFilter

+++ app/src/main/java/com/hifitoy/hifitoyobjects/HiFiToyObject.java
@@ -16,7 +16,7 @@ import java.io.IOException;
 
 public interface HiFiToyObject {
 
-    int getAddress();
+    byte getAddress();
     String getInfo();
 
     void sendToPeripheral(boolean response);
